Fix uploading when the path to the uploaded files is the shared container
resolvingSymlinksInPath() only works if the file exists so for the temp file created as the path that the archive is copied to, that will always include “/private” and therefore will break when compared against the path to the base directory (which does exist). While this is fix should be revisited, this will get it working for the app group identifier case (which I haven’t figured out how to set up testing for yet).